flask-backend/requirements.txt
flask==3.1.0
flask-cors==6.0.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.30
python-dotenv==1.0.1